requirements.txt
playwright==1.18.2
requests==2.27.1
python-dotenv==0.19.2